AsteroidSpawns
Added quick-n-dirty sprite for asteroids.
Added Asteroid Spawn script to main camera.
Made Asteroids spawn at a random point on a circle outside the viewport.
When the asteroid spawns, they are directed at a random point on the
screen and given a random push in that direction.
Currently they do not despawn. Probably gonna put something in the
update for checking position.