Created a shooting mode
Script goes on the player object. When the player object is clicked, the
game enters 'Shooting mode'. The physics engine is slowed to a stop, the
player object rotates to face the location of the mouse.
If the player clicks on the player again, shooting mode is deactivated,
the physics engine speeds up again.
I included code that should work for touch devices. However I have no
way of testing it...